Update dependencies and refactor logging module
- Bump `coverage[toml]` to 7.6.1 in `requirements-dev.txt`
- Bump `ruff` to 0.5.6 in `requirements-dev.txt` and `.pre-commit-config.yaml`
- Bump `wheel` to 0.44.0 in `requirements-dev.txt`
- Bump `hcloud` to 2.2.0 in `requirements.txt`
- Bump `pyyaml` to 6.0.2 in `requirements.txt`
- Refactor logging module import to `custom_logging` across multiple files
